# be/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "PlanWise API Documentation",
        "description": "API Documentation for PlanWise application - a smart daily planning application with AI features.",
        "contact": {
            "name": "PlanWise Team",
            "email": "contact@planwise.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://springdoc.org"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Local development"
        }
    ],
    "security": [
        {
            "bearer": []
        }
    ],
    "paths": {
...
        "/api/v1/habits/{habitId}": {
            "get": {
                "tags": [
                    "habit-controller"
                ],
                "operationId": "getHabitById",
                "parameters": [
                    {
                        "name": "habitId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                        },
                        "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/HabitDto"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "habit-controller"
                ],
                "operationId": "updateHabit",
                "parameters": [
                    {
                        "name": "habitId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                        },
                        "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateHabitRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/HabitDto"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "habit-controller"
                ],
                "operationId": "deleteHabit",
                "parameters": [
                    {
                        "name": "habitId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                        },
                        "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/api/v1/habits": {
            "get": {
                "tags": [
                    "habit-controller"
                ],
                "operationId": "getHabits",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/HabitListResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "habit-controller"
                ],
                "operationId": "createHabit",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateHabitRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/HabitDto"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/notifications": {
            "get": {
                "tags": [
                    "notification-controller"
                ],
                "operationId": "getNotifications",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/NotificationDto"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "notification-controller"
                ],
                "operationId": "updateAllNotifications",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateNotificationRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/v1/notifications/{notificationId}": {
            "delete": {
                "tags": [
                    "notification-controller"
                ],
                "operationId": "deleteNotification",
                "parameters": [
                    {
                        "name": "notificationId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                        },
                        "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "patch": {
                "tags": [
                    "notification-controller"
                ],
                "operationId": "updateNotification",
                "parameters": [
                    {
                        "name": "notificationId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                        },
                        "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateNotificationRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotificationDto"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/habits/{habitId}/completions/{date}": {
            "patch": {
                "tags": [
                    "habit-controller"
                ],
                "operationId": "toggleCompletion",
                "parameters": [
                    {
                        "name": "habitId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                        },
                        "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                    },
                    {
                        "name": "date",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2026-06-05"
                        },
                        "example": "2026-06-05"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/HabitDto"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UpdateHabitRequest": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string",
                        "maxLength": 255,
                        "minLength": 0
                    },
                    "description": {
                        "type": "string"
                    },
                    "frequency": {
                        "type": "string",
                        "enum": [
                            "daily",
                            "weekly",
                            "monthly"
                        ],
                        "example": "daily"
                    },
                    "targetCount": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "color": {
                        "type": "string",
                        "example": "indigo"
                    },
                    "isActive": {
                        "type": "boolean"
                    },
                    "sortOrder": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
            "HabitDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "frequency": {
                        "type": "string",
                        "enum": [
                            "daily",
                            "weekly",
                            "monthly"
                        ],
                        "example": "daily"
                    },
                    "targetCount": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "currentStreak": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "bestStreak": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "color": {
                        "type": "string",
                        "example": "indigo"
                    },
                    "isActive": {
                        "type": "boolean"
                    },
                    "sortOrder": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "completedDates": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "date",
                            "example": "2026-06-05"
                        },
                        "uniqueItems": true
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-06-05T08:00:00.000Z"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-06-05T08:00:00.000Z"
                    }
                }
            },
...
            "CreateHabitRequest": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string",
                        "maxLength": 255,
                        "minLength": 0
                    },
                    "description": {
                        "type": "string"
                    },
                    "frequency": {
                        "type": "string",
                        "enum": [
                            "daily",
                            "weekly",
                            "monthly"
                        ],
                        "example": "daily"
                    },
                    "targetCount": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "color": {
                        "type": "string",
                        "example": "indigo"
                    }
                },
                "required": [
                    "title"
                ]
            },
...
            "UpdateNotificationRequest": {
                "type": "object",
                "properties": {
                    "read": {
                        "type": "boolean"
                    },
                    "dismissed": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "dismissed",
                    "read"
                ]
            },
            "NotificationDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                    },
                    "type": {
                        "type": "string"
                    },
                    "tone": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    },
                    "ctaLabel": {
                        "type": "string"
                    },
                    "read": {
                        "type": "boolean"
                    },
                    "dismissed": {
                        "type": "boolean"
                    },
                    "scheduledFor": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-06-05T08:00:00.000Z"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-06-05T08:00:00.000Z"
                    }
                }
            },
...
            "HabitListResponse": {
                "type": "object",
                "properties": {
                    "habits": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/HabitDto"
                        }
                    },
                    "totalCount": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "activeCount": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "isPremium": {
                        "type": "boolean"
                    },
                    "freeLimit": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Enter your Bearer JWT token to access protected endpoints.",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}